example/requirements.txt
asynction
gevent-websocket==0.10.1
typing-extensions==3.10.0.0
